backend/app/requirements.txt
amqp==2.5.2
asgiref==3.2.3
attrs==19.3.0
billiard==3.6.3.0
celery==4.4.2
celery-once==3.0.1
certifi==2019.11.28
click==7.0
coverage==5.1
Django==3.0.3
django-click==2.1.0
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-gis==0.15
djangorestframework-simplejwt==4.4.0
entrypoints==0.3
factory-boy==2.12.0
Faker==4.0.3
flake8==3.7.9
importlib-metadata==1.5.0
isort==4.3.21
kombu==4.6.8
Markdown==3.2.1
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.1
pytest-cov==2.8.1
pytest-django==3.9.0
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
raven==6.10.0
redis==3.4.1
sentry-sdk==0.14.2
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.9
zipp==3.1.0
